requirements.txt
tweepy == 3.10.0
python-dotenv == 0.15.0
